requirements.txt
kafka-python 
celery 
redis 
requests
elasticsearch 
beautifulsoup4
apache-airflow
fastapi
uvicorn 
